+++ include/gtest/iutest_gmock_ver.hpp
@@ -15,12 +15,6 @@
 #ifndef INCG_IRIS_IUTEST_GMOCK_VER_HPP_CBBF82C8_EB6F_4398_BAA6_5B485AC52D36_
 #define INCG_IRIS_IUTEST_GMOCK_VER_HPP_CBBF82C8_EB6F_4398_BAA6_5B485AC52D36_
 
-#if defined(IUTEST_USE_GTEST) && defined(IUTEST_USE_GMOCK)
-
-//======================================================================
-// include
-#include <gmock/gmock.h>
-
 //======================================================================
 // define
 
@@ -32,9 +26,11 @@
 
 //!< Minor Version
 #ifndef GMOCK_MINORVER
-#  if   defined(GMOCK_INCLUDE_GMOCK_INTERNAL_CUSTOM_GMOCK_MATCHERS_H_)
+#  if   !defined(GTEST_INCLUDE_GTEST_INTERNAL_GTEST_LINKED_PTR_H_)
 #    define GMOCK_MINORVER  0x08
 #    define GMOCK_LATEST    1
+#  elif defined(GMOCK_INCLUDE_GMOCK_INTERNAL_CUSTOM_GMOCK_MATCHERS_H_)
+#    define GMOCK_MINORVER  0x08
 #  elif defined(GMOCK_INCLUDE_GMOCK_INTERNAL_CUSTOM_CALLBACK_MATCHERS_H_)
 #    define GMOCK_MINORVER  0x08
 #  elif defined(GMOCK_GMOCK_MORE_MATCHERS_H_)
@@ -50,8 +46,19 @@
 #  endif
 #endif
 
+//!< Micro Version
+#ifndef GMOCK_MICROVER
+#  if   GMOCK_MINORVER == 0x08 && !defined(GMOCK_INCLUDE_GMOCK_INTERNAL_CUSTOM_CALLBACK_MATCHERS_H_)
+#    define GMOCK_MICROVER  0x01
+#  endif
+#endif
+
+#ifndef GMOCK_MICROVER
+#  define GMOCK_MICROVER  0x00
+#endif
+
 #ifndef GMOCK_BUILD
-#  define GMOCK_BUILD       0x00        //!< Build
+#  define GMOCK_BUILD       GMOCK_MICROVER  //!< @deprecated
 #endif
 
 #ifndef GMOCK_REVISION
@@ -61,15 +68,13 @@
 #ifndef GMOCK_VER
 #  define   GMOCK_VER       ((GMOCK_MAJORVER<<24)   \
                             |(GMOCK_MINORVER<<16)   \
-                            |(GMOCK_BUILD<<8)       \
+                            |(GMOCK_MICROVER<<8)    \
                             |(GMOCK_REVISION<<0)    \
-                            )   //!< google test version
+                            )   //!< google mock version
 #endif
 
 #if !defined(GMOCK_LATEST)
-#  define GMOCK_LATEST      0   //!< google test development latest version
+#  define GMOCK_LATEST      0   //!< google mock development latest version
 #endif
 
 #endif
-
-#endif // INCG_IRIS_IUTEST_GMOCK_VER_HPP_CBBF82C8_EB6F_4398_BAA6_5B485AC52D36_