+++ include/gtest/iutest_gtest_ver.hpp
@@ -15,11 +15,6 @@
 #ifndef INCG_IRIS_IUTEST_GTEST_VER_HPP_7396AEF4_D1AB_44B5_9FE3_23CEC1240158_
 #define INCG_IRIS_IUTEST_GTEST_VER_HPP_7396AEF4_D1AB_44B5_9FE3_23CEC1240158_
 
-#if defined(IUTEST_USE_GTEST)
-
-//======================================================================
-// include
-#include <gtest/gtest.h>
 
 //======================================================================
 // define
@@ -32,7 +27,7 @@
 
 //!< Minor Version
 #ifndef GTEST_MINORVER
-#  if   defined(GTEST_STRINGIFY_)
+#  if   !defined(GTEST_INCLUDE_GTEST_INTERNAL_GTEST_LINKED_PTR_H_)
 #    define GTEST_MINORVER  0x08
 #    define GTEST_LATEST    1
 #  elif defined(GTEST_ATTRIBUTE_NO_SANITIZE_ADDRESS_)
@@ -58,14 +53,18 @@
 
 //!< Micro Version
 #ifndef GTEST_MICROVER
-#  if GTEST_MINORVER == 0x02 && defined(GTEST_HAS_CLONE)
+#  if   GTEST_MINORVER == 0x08 && defined(GTEST_STRINGIFY_)
+#    define GTEST_MICROVER    0x01
+#  elif GTEST_MINORVER == 0x02 && defined(GTEST_HAS_CLONE)
 #    define GTEST_MICROVER  0x01
 #  elif GTEST_MINORVER == 0x00
 // 判断難しい
 #    define GTEST_MICROVER  0x01
-#  else
-#    define GTEST_MICROVER  0x00
 #  endif
+#endif
+
+#ifndef GTEST_MICROVER
+#  define GTEST_MICROVER    0x00
 #endif
 
 #ifndef GTEST_BUILD
@@ -89,5 +88,3 @@
 #endif
 
 #endif
-
-#endif // INCG_IRIS_IUTEST_GTEST_VER_HPP_7396AEF4_D1AB_44B5_9FE3_23CEC1240158_