+++ include/gtest/switch/iutest_switch_cmphelper.hpp
@@ -6,7 +6,7 @@
  *
  * @author      t.shirayanagi
  * @par         copyright
- * Copyright (C) 2012-2016, Takazumi Shirayanagi\n
+ * Copyright (C) 2012-2019, Takazumi Shirayanagi\n
  * This software is released under the new BSD License,
  * see LICENSE
 */
@@ -27,6 +27,10 @@ namespace testing
 
 namespace internal
 {
+
+#if GTEST_VER > 0x01080100 || GTEST_LATEST
+using ::std::is_pointer;
+#endif
 
 #if GTEST_MINORVER < 0x06
 template<bool> struct EnableIf;