+++ include/listener/iutest_junit_xml_generator.hpp
@@ -6,7 +6,7 @@
  *
  * @author      t.shirayanagi
  * @par         copyright
- * Copyright (C) 2014-2016, Takazumi Shirayanagi\n
+ * Copyright (C) 2014-2018, Takazumi Shirayanagi\n
  * This software is released under the new BSD License,
  * see LICENSE
 */
@@ -42,6 +42,7 @@ public:
     {
     }
 public:
+    virtual bool IsReportable(const UnitTest& test) IUTEST_CXX_OVERRIDE;
     virtual void OnReportTest(IFile* file, const UnitTest& test) IUTEST_CXX_OVERRIDE;
 
 private: