+++ test/unit_tests.cpp
@@ -203,11 +203,27 @@ IUTEST(UnitTest, FormatSizeByte)
 {
     IUTEST_EXPECT_STREQ("0B", ::iutest::detail::FormatSizeByte(0));
     IUTEST_EXPECT_STREQ("2B", ::iutest::detail::FormatSizeByte(2));
+}
+
+IUTEST(UnitTest, FormatSizeKByte)
+{
     IUTEST_EXPECT_STREQ("1KB", ::iutest::detail::FormatSizeByte(1024));
     IUTEST_EXPECT_STREQ("1.0KB", ::iutest::detail::FormatSizeByte(1025));
+}
+
+IUTEST(UnitTest, FormatSizeMByte)
+{
     IUTEST_EXPECT_STREQ("1MB", ::iutest::detail::FormatSizeByte(1024 * 1024));
     IUTEST_EXPECT_STREQ("1.9MB", ::iutest::detail::FormatSizeByte(2 * 1024 * 1024 - 1));
+}
+
+IUTEST(UnitTest, FormatSizeGByte)
+{
     IUTEST_EXPECT_STREQ("1GB", ::iutest::detail::FormatSizeByte(1024 * 1024 * 1024));
+}
+
+IUTEST(UnitTest, FormatSizeTByte)
+{
     IUTEST_EXPECT_STREQ("1TB", ::iutest::detail::FormatSizeByte(1024ull * 1024 * 1024 * 1024));
     IUTEST_EXPECT_STREQ("1024TB", ::iutest::detail::FormatSizeByte(1024ull * 1024 * 1024 * 1024 * 1024));
 }